moved symf wrapper to own file
