+++ vscode/src/chat/chat-view/ChatController.ts
@@ -78,7 +78,7 @@ import { isRunningInsideAgent } from '../../jsonrpc/isRunningInsideAgent'
 import { ContextStatusAggregator } from '../../local-context/enhanced-context-status'
 import type { LocalEmbeddingsController } from '../../local-context/local-embeddings'
 import { rewriteChatQuery } from '../../local-context/rewrite-chat-query'
-import type { SymfWrapper } from '../../local-context/symf'
+import type { SymfWrapper } from '../../local-context/symf/symf-wrapper'
 import { logDebug } from '../../log'
 import { migrateAndNotifyForOutdatedModels } from '../../models/modelMigrator'
 import { gitCommitIdFromGitExtension } from '../../repository/git-extension-api'
