+++ vscode/src/chat/chat-view/context.ts
@@ -25,7 +25,8 @@ import type { RemoteSearch } from '../../context/remote-search'
 import { resolveContextItems } from '../../editor/utils/editor-context'
 import type { VSCodeEditor } from '../../editor/vscode-editor'
 import type { LocalEmbeddingsController } from '../../local-context/local-embeddings'
-import type { SymfRunner, SymfWrapper } from '../../local-context/symf'
+import type { SymfRunner } from '../../local-context/symf/symf-runner'
+import type { SymfWrapper } from '../../local-context/symf/symf-wrapper'
 import { logDebug, logError } from '../../log'
 import { repoNameResolver } from '../../repository/repo-name-resolver'
