+++ vscode/src/local-context/symf/download-symf.ts
@@ -3,11 +3,11 @@ import os from 'node:os'
 import path from 'node:path'
 import type { SemverString } from '@sourcegraph/cody-shared/src/utils'
 import * as vscode from 'vscode'
-import { waitForLock } from '../lockfile'
-import { logDebug, logError } from '../log'
-import { type Arch, Platform, getOSArch } from '../os'
-import { captureException } from '../services/sentry/sentry'
-import { downloadFile, fileExists, unzip } from './utils'
+import { waitForLock } from '../../lockfile'
+import { logDebug, logError } from '../../log'
+import { type Arch, Platform, getOSArch } from '../../os'
+import { captureException } from '../../services/sentry/sentry'
+import { downloadFile, fileExists, unzip } from '../utils'
 
 export type SymfVersionString = SemverString<'v'>
 const symfVersion: SymfVersionString = 'v0.0.13'
