+++ vscode/src/main.ts
@@ -56,7 +56,7 @@ import type { PlatformContext } from './extension.common'
 import { configureExternalServices } from './external-services'
 import { isRunningInsideAgent } from './jsonrpc/isRunningInsideAgent'
 import type { LocalEmbeddingsController } from './local-context/local-embeddings'
-import type { SymfWrapper } from './local-context/symf'
+import type { SymfWrapper } from './local-context/symf/symf-wrapper'
 import { logDebug, logError } from './log'
 import { MinionOrchestrator } from './minion/MinionOrchestrator'
 import { PoorMansBash } from './minion/environment'
